refactor(tui-components): separate motion rendering from playback policy
Accept explicit scene, time, and transition progress; retain MotionState as an optional controller returning owned frames. Split palette, sampling, zoom, tile geometry, and state boundaries. Preserve existing visual snapshots and document the API migration and retained profiling evidence.
